Removed code from ClientHandler and TRServerMain
A lot of the code in this file is only relevant to the context which I lifted it from (a homework assignment from a previous class) and isn't needed here. Hence it has been removed.

Some minor adjustments too, like changing the thread pool we use from fixed to cached.